requirements.txt
celery==4.3.0
Django==3.0
django-heroku==0.3.1
gunicorn==20.0.4
redis==3.3.11
SQLAlchemy==1.3.11
django-celery-beat==1.5.0
django-cleanup==4.0.0
Pillow==6.2.1
django-simple-history==2.8.0
